jQuery on "New Father" & Changing "New User" Form
Now on the "New Father" form when we create a new father we have to complete
if the father has children attending at our school and if the student is
going to a school of our network.

The "New User" form is changed the view and the controller. The outcome is
the same as before, but now in the controller we have SQLFORM.factory from
the db, and we manipulate the form in the view.