Updating "New Father" Form
Now the "New Father" Form is customized. It's benn added the JS
to conditional fields for "priority fathers". For this, there is
a new table on the DB, 'priority_father' in where we store if the
father has some children attending at our school and/or if the
student is comming from one of the schools in our network.