New Dates Form is using two forms.
The first form selects the year and with this retrieves the data from the DB.
The second form is the one that populates the DB.
TODO: The second form needs validation! (via a validator or jQuery).
The selected_year field is writable and I don't like it.
Perhaps it's a good idea to hide the Field or make it non-writable and non-readable (I don't know how yet).